requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
